# openapi.yaml
openapi: 3.1.0
info:
  title: Open Arena API
  version: 0.3.0
  summary: Leaderboard-scoped, environment-centric REST API for Open Arena
  description: |
    Open Arena is modeled around leaderboards and environments.

    A leaderboard defines the comparison perimeter: one scoped model catalog plus
    one or more environments. An environment is the atomic evaluation artifact:
    benchmark workload + inline or referenced dataset definition + verifier
    ensemble + optional sandbox/reset/runtime policies.

    The API exposes the Arena's two execution modes explicitly:
      1. generator: direct model execution over the environment workload
      2. agent: tool-enabled or remote-agent execution with trajectory capture

    Leaderboards can be created in one shot from a full config-like payload, or
    created empty and populated later through nested model-catalog and
    environment membership endpoints.

    Runs may target:
      - an entire leaderboard
      - one or more environments inside a leaderboard across all leaderboard models
      - one or more models inside a leaderboard across all leaderboard environments
      - explicit model × environment slices supplied directly in the run payload.

    In this artifact, `Run` is scoped to evaluation submissions plus their
    cache-backed result materialization. Job-control primitives (cancel/retry/
    progress/worker leases) and training or optimization jobs are intentionally
...

...
servers:
...
security:
  - bearerAuth: []
paths:
  /v1/verifiers:
    get:
      tags: [Verifiers]
      summary: List reusable verifier suites
      operationId: listVerifiers
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Verifier suites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifierSuiteListResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Verifiers]
      summary: Create reusable verifier suite
      operationId: createVerifierSuite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifierSuiteCreate'
      responses:
        '201':
          description: Verifier suite created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifierSuite'
        default:
          $ref: '#/components/responses/Error'
  /v1/verifiers/{verifier_id}:
    get:
      tags: [Verifiers]
      summary: Get verifier suite
      operationId: getVerifierSuite
      parameters:
        - $ref: '#/components/parameters/VerifierId'
      responses:
        '200':
          description: Verifier suite
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifierSuite'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Verifiers]
      summary: Update verifier suite
      operationId: updateVerifierSuite
      parameters:
        - $ref: '#/components/parameters/VerifierId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifierSuitePatch'
      responses:
        '200':
          description: Updated verifier suite
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifierSuite'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Verifiers]
      summary: Delete verifier suite
      operationId: deleteVerifierSuite
      parameters:
        - $ref: '#/components/parameters/VerifierId'
      responses:
        '204':
          description: Verifier suite deleted
        default:
          $ref: '#/components/responses/Error'
  /v1/environments:
    get:
      tags: [Environments]
      summary: List reusable environments
      operationId: listEnvironments
      parameters:
        - in: query
          name: source_kind
          schema:
            $ref: '#/components/schemas/EnvironmentSourceKind'
        - in: query
          name: mode
          schema:
            $ref: '#/components/schemas/RunMode'
        - in: query
          name: name
          schema:
            type: string
          description: Lookup by canonical environment name.
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Environment registry entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentListResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Environments]
      summary: Create reusable environment
      operationId: createEnvironment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentCreate'
      responses:
        '201':
          description: Environment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Environment'
        default:
          $ref: '#/components/responses/Error'
  /v1/environments/{environment_id}:
    get:
      tags: [Environments]
      summary: Get reusable environment
      operationId: getEnvironment
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
      responses:
        '200':
          description: Environment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Environment'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Environments]
      summary: Update reusable environment
      operationId: updateEnvironment
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentPatch'
      responses:
        '200':
          description: Updated environment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Environment'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Environments]
      summary: Delete reusable environment
      operationId: deleteEnvironment
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
      responses:
        '204':
          description: Environment deleted
        default:
          $ref: '#/components/responses/Error'
  /v1/environments/{environment_id}/versions:
    get:
      tags: [Environments]
      summary: List available versions for an environment
      description: |
        Returns the known version descriptors for this environment.
        For git-backed sources (github_repo, prime_environment_hub) the
        currently-resolved commit is always included; when the backend can
        enumerate tags/refs additional entries may appear.
        For inline environments a single entry derived from the source
        version is returned.
      operationId: listEnvironmentVersions
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
      responses:
        '200':
          description: Environment version list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentVersionListResponse'
        default:
          $ref: '#/components/responses/Error'
  /v1/leaderboards:
    get:
      tags: [Leaderboards]
      summary: List leaderboards
      operationId: listLeaderboards
      parameters:
        - in: query
          name: visibility
          schema:
            $ref: '#/components/schemas/LeaderboardVisibility'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Leaderboards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardListResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Leaderboards]
      summary: Create leaderboard
      description: |
        Create either:
          - a full leaderboard with embedded model catalog + environments, or
          - an empty leaderboard shell that can be populated later.
      operationId: createLeaderboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaderboardCreate'
      responses:
        '201':
          description: Leaderboard created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Leaderboard'
        default:
          $ref: '#/components/responses/Error'
  /v1/leaderboards/{leaderboard_id}:
    get:
      tags: [Leaderboards]
      summary: Get leaderboard
      operationId: getLeaderboard
      parameters:
        - $ref: '#/components/parameters/LeaderboardId'
      responses:
        '200':
          description: Leaderboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Leaderboard'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Leaderboards]
      summary: Update leaderboard metadata or ranking policy
      operationId: updateLeaderboard
      parameters:
        - $ref: '#/components/parameters/LeaderboardId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaderboardPatch'
      responses:
        '200':
          description: Updated leaderboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Leaderboard'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Leaderboards]
      summary: Delete leaderboard
      operationId: deleteLeaderboard
      parameters:
        - $ref: '#/components/parameters/LeaderboardId'
      responses:
        '204':
          description: Leaderboard deleted
        default:
          $ref: '#/components/responses/Error'
  /v1/leaderboards/{leaderboard_id}/model-catalog:
    get:
      tags: [Leaderboards]
      summary: Get leaderboard-scoped model catalog
      operationId: getLeaderboardModelCatalog
      parameters:
        - $ref: '#/components/parameters/LeaderboardId'
      responses:
        '200':
          description: Leaderboard model catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelCatalog'
        default:
          $ref: '#/components/responses/Error'
    put:
      tags: [Leaderboards]
      summary: Replace leaderboard model catalog
      operationId: replaceLeaderboardModelCatalog
      parameters:
        - $ref: '#/components/parameters/LeaderboardId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelCatalogReplace'
      responses:
        '200':
          description: Replaced model catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelCatalog'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Leaderboards]
      summary: Patch leaderboard model catalog metadata
      operationId: patchLeaderboardModelCatalog
      parameters:
        - $ref: '#/components/parameters/LeaderboardId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelCatalogPatch'
      responses:
        '200':
          description: Updated model catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelCatalog'
        default:
          $ref: '#/components/responses/Error'
  /v1/leaderboards/{leaderboard_id}/models:
    get:
      tags: [Leaderboards]
      summary: List models in a leaderboard model catalog
      operationId: listLeaderboardModels
      parameters:
        - $ref: '#/components/parameters/LeaderboardId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Models
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelDefinitionListResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Leaderboards]
      summary: Add model to leaderboard model catalog
      operationId: addLeaderboardModel
      parameters:
        - $ref: '#/components/parameters/LeaderboardId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelDefinitionCreate'
      responses:
        '201':
          description: Model added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelDefinition'
        default:
          $ref: '#/components/responses/Error'
  /v1/leaderboards/{leaderboard_id}/models/{model_id}:
    get:
      tags: [Leaderboards]
      summary: Get model in leaderboard model catalog
      operationId: getLeaderboardModel
      parameters:
        - $ref: '#/components/parameters/LeaderboardId'
        - $ref: '#/components/parameters/ModelId'
      responses:
        '200':
          description: Model definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelDefinition'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Leaderboards]
      summary: Update model in leaderboard model catalog
      operationId: updateLeaderboardModel
      parameters:
        - $ref: '#/components/parameters/LeaderboardId'
        - $ref: '#/components/parameters/ModelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelDefinitionPatch'
      responses:
        '200':
          description: Updated model definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelDefinition'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Leaderboards]
      summary: Remove model from leaderboard model catalog
      operationId: deleteLeaderboardModel
      parameters:
        - $ref: '#/components/parameters/LeaderboardId'
        - $ref: '#/components/parameters/ModelId'
      responses:
        '204':
          description: Model removed
        default:
          $ref: '#/components/responses/Error'
  /v1/leaderboards/{leaderboard_id}/environments:
    get:
      tags: [Leaderboards]
      summary: List environments in leaderboard scope
      operationId: listLeaderboardEnvironments
      parameters:
        - $ref: '#/components/parameters/LeaderboardId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Leaderboard environments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentMembershipListResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Leaderboards]
      summary: Add environment to leaderboard scope
      operationId: addLeaderboardEnvironment
      parameters:
        - $ref: '#/components/parameters/LeaderboardId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentMembershipCreate'
      responses:
        '201':
          description: Environment attached to leaderboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentMembership'
        default:
          $ref: '#/components/responses/Error'
  /v1/leaderboards/{leaderboard_id}/environments/{environment_id}:
    get:
      tags: [Leaderboards]
      summary: Get leaderboard environment membership
      operationId: getLeaderboardEnvironment
      parameters:
        - $ref: '#/components/parameters/LeaderboardId'
        - $ref: '#/components/parameters/EnvironmentId'
      responses:
        '200':
          description: Environment membership
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentMembership'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Leaderboards]
      summary: Update environment membership overrides
      operationId: updateLeaderboardEnvironment
      parameters:
        - $ref: '#/components/parameters/LeaderboardId'
        - $ref: '#/components/parameters/EnvironmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentMembershipPatch'
      responses:
        '200':
          description: Updated environment membership
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentMembership'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Leaderboards]
      summary: Remove environment from leaderboard scope
      operationId: deleteLeaderboardEnvironment
      parameters:
        - $ref: '#/components/parameters/LeaderboardId'
        - $ref: '#/components/parameters/EnvironmentId'
      responses:
        '204':
          description: Environment removed
        default:
          $ref: '#/components/responses/Error'
  /v1/leaderboards/{leaderboard_id}/entries:
    get:
      tags: [Leaderboards]
      summary: List leaderboard entries
      operationId: listLeaderboardEntries
      parameters:
        - $ref: '#/components/parameters/LeaderboardId'
        - in: query
          name: environment_id
          schema:
            type: string
            format: uuid
        - in: query
          name: model_id
          schema:
            type: string
            format: uuid
        - in: query
          name: as_of
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Leaderboard entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardEntryListResponse'
        default:
          $ref: '#/components/responses/Error'
  /v1/metric-kinds:
    get:
      tags: [Discovery]
      summary: List discovered metric kinds
      description: |
        Lists the reward and metric identifiers this server can resolve in
        `MetricDefinition.metric_kind`. These identifiers are auto-discovered
        from the server's reward/metric registry and are distinct from the
        user-managed verifier suite resources.
      operationId: listMetricKinds
      responses:
        '200':
          description: Supported metric and reward identifiers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryIdentifierListResponse'
        default:
          $ref: '#/components/responses/Error'
  /v1/aggregations:
    get:
      tags: [Discovery]
      summary: List discovered aggregation identifiers
      description: |
        Lists aggregation identifiers supported by this server for verifier
        suites, ranking policies, and result aggregates. Use the optional
        `scopes` metadata on each item to distinguish where a given identifier
        applies. Every compliant server must expose `weighted_mean` as a
        baseline identifier because it is the default aggregation in this API.
      operationId: listAggregations
      responses:
        '200':
          description: Supported aggregation identifiers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryIdentifierListResponse'
        default:
          $ref: '#/components/responses/Error'
  /v1/model-providers:
    get:
      tags: [Discovery]
      summary: List discovered model providers
      description: |
        Lists provider identifiers that can be used in
        `ModelExecutionConfig.provider`.
      operationId: listModelProviders
      responses:
        '200':
          description: Supported model provider identifiers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryIdentifierListResponse'
        default:
          $ref: '#/components/responses/Error'
  /v1/dataset-providers:
    get:
      tags: [Discovery]
      summary: List discovered dataset providers
      description: |
        Lists provider identifiers that can be used in
        `DatasetBinding.provider`.
      operationId: listDatasetProviders
      responses:
        '200':
          description: Supported dataset provider identifiers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryIdentifierListResponse'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      in: query
      name: cursor
      schema:
        type: string
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  schemas:
...
    RunMode:
      type: string
      enum: [generator, agent]
...